chatbot_backend/requirements-base.txt
# Core Web Framework
fastapi==0.120.1
uvicorn==0.38.0
starlette==0.49.1
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.1.0

# Database
motor==3.7.1
pymongo==4.15.3

# HTTP & Networking
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
sniffio==1.3.1

# Document Processing
pdf2image==1.17.0
pillow==12.0.0
pytesseract==0.3.13
python-docx==0.8.11
PyPDF2==3.0.1

# Google Services
google-auth==2.42.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.106.0
google-generativeai==0.8.5

# Web Scraping
playwright==1.42.0
pandas==2.3.3
selectolax==0.4.0

# Utilities
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
click==8.3.0
tenacity==9.1.2
tqdm==4.67.1
dnspython==2.8.0

# System
typing_extensions==4.15.0
anyio==4.11.0
